requirements.txt
Django >=5.2, <5.3
djangorestframework>=3.16.1, <3.17
psycopg[binary]>=3.1,<4.0
drf-spectacular>=0.28.0,<0.29.0
djangorestframework-simplejwt==5.3.1
django-cors-headers>=4.3.0

